requirements.txt
flask==3.0.0
google-genai
gunicorn==21.2.0
werkzeug==3.0.1
python-dotenv==1.0.0
pytest==7.4.3
flask-compress
google-cloud-logging
firebase-admin
google-cloud-storage
pydantic
google-cloud-aiplatform
